hotfix(contact): solve button in contact
